% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxNCSpline.R
\name{CoxNCSpline}
\alias{CoxNCSpline}
\title{This function implments natural cubic splines in Cox regression for latency analysis}
\usage{
CoxNCSpline(
  data,
  time_start,
  time_end,
  status,
  exposure,
  knots_number,
  latency
)
}
\arguments{
\item{data}{A data frame containing the data}

\item{time_start}{start time}

\item{time_end}{end time}

\item{status}{survival status. 1 for death and 0 for censored}

\item{exposure}{Name of the time-varying exposure variable. From the most recent to the furthest in time.}

\item{knots_number}{Pre-specified number of knots}

\item{latency}{prespecified latency}
}
\value{
A list. The first element is the best knots with the lowest AIC. The second element is a Cox model
}
\description{
This function implments natural cubic splines in Cox regression for latency analysis
}
\examples{
time_start <- "age_start"
time_end <- "age_end"
status <- "failure"
exposure <- paste0("lag", 0:19)
knots_number <- 3
latency <- 20
fit <- CoxNCSpline(sim_data, time_start, time_end, status, exposure, knots_number, latency)
}
