% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxPoly.R
\name{CoxPoly}
\alias{CoxPoly}
\title{This is a function to implement polynomial Cox regression for latency analysis}
\usage{
CoxPoly(
  data,
  time_start,
  time_end,
  status,
  exposure,
  degree,
  latency,
  adjusted_variable = NULL,
  adjusted_model = NULL,
  id = "id"
)
}
\arguments{
\item{data}{A data frame containing the data}

\item{time_start}{start time}

\item{time_end}{end time}

\item{status}{survival status. 1 for death and 0 for censored}

\item{exposure}{Name of the time-varying exposure variable. From the most recent to the furthest in time.}

\item{degree}{degree of the polynomial}

\item{latency}{prespecified latency}

\item{adjusted_variable}{A vector of adjusted variables}

\item{adjusted_model}{A vector of adjusted models}

\item{id}{A string to specify the name of the id column}
}
\value{
A Cox model object
}
\description{
This is a function to implement polynomial Cox regression for latency analysis
}
\examples{
time_start <- "age_start"
time_end <- "age_end"
status <- "failure"
exposure <- paste0("lag", 0:19)
degree <- 3
latency <- 20
fit <- CoxPoly(sim_data, time_start, time_end, status, exposure, degree, latency)
fit
}
