% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxKnotSearch.R
\name{extract_CoxKnotsearch}
\alias{extract_CoxKnotsearch}
\title{This is a function to extract log HR based on the resuls from CoxKnotsearch}
\usage{
extract_CoxKnotsearch(fit, lag, latency)
}
\arguments{
\item{fit}{A model from CoxKnotsearch}

\item{lag}{A value of lag time}

\item{latency}{prespecified latency}
}
\value{
The log HR estimated at that lag time
}
\description{
This is a function to extract log HR based on the resuls from CoxKnotsearch
}
\examples{
time_start <- "age_start"
time_end <- "age_end"
status <- "failure"
exposure <- paste0("lag", 0:19)
# knots can be set as as a vector of potential number of knots, here we use 3 to save time
knots_number <- c(3)
latency <- 20
fit <- CoxKnotsearch(sim_data, time_start, time_end, status, exposure, knots_number, latency)
extract_CoxKnotsearch(fit, 10, latency = latency)
}
