% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_clusters.R
\name{resample_clusters}
\alias{resample_clusters}
\title{This is a function to help resemple clusters for bootstrap. It is adapted from the function "clusterboot" in the R package "fpc".}
\usage{
resample_clusters(data, id = "id")
}
\arguments{
\item{data}{A data frame containing the data}

\item{id}{A string of the column name that contains the subject ID}
}
\value{
A data frame with the same structure as the input data, but with resampled clusters
}
\description{
This is a function to help resemple clusters for bootstrap. It is adapted from the function "clusterboot" in the R package "fpc".
}
