% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{example
A simulated data for the use of example}
\format{
An object of class \code{data.frame} with 9246 rows and 38 columns.
}
\usage{
sim_data
}
\description{
example
A simulated data for the use of example
}
\keyword{datasets}
