% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnotSearch.R
\name{CoxKnotsearch}
\alias{CoxKnotsearch}
\title{This function implments GridSearch of natural cubic splines in Cox regression for latency analysis
This is GreadSearching by knots}
\usage{
CoxKnotsearch(
  data,
  time_start,
  time_end,
  status,
  exposure,
  knots_number,
  latency,
  adjusted_variable = NULL,
  adjusted_model = NULL
)
}
\arguments{
\item{data}{A data frame containing the data}

\item{time_start}{start time}

\item{time_end}{end time}

\item{status}{survival status. 1 for death and 0 for censored}

\item{exposure}{Name of the time-varying exposure variable. From the most recent to the furthest in time.}

\item{knots_number}{A vector of potential number of knots}

\item{latency}{pre-specified latency}
}
\value{
A list. The first element is the best knots with the lowest AIC. The second element is a Cox model
}
\description{
This function implments GridSearch of natural cubic splines in Cox regression for latency analysis
This is GreadSearching by knots
}
\examples{
time_start <- "age_start"
time_end <- "age_end"
status <- "failure"
exposure <- paste0("lag", 0:19)
knots_number <- c(3,4,5)
latency <- 20
fit <- CoxKnotsearch(sim_data, time_start, time_end, status, exposure, knots_number, latency)
}
