% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxKnotSearch.R
\name{CoxKnotsearch_boot}
\alias{CoxKnotsearch_boot}
\title{This function conduct bootstraps for @CoxKnotsearch}
\usage{
CoxKnotsearch_boot(
  data,
  time_start,
  time_end,
  status,
  exposure,
  knots_number,
  latency,
  adjusted_variable = NULL,
  adjusted_model = NULL,
  lag,
  parallel = FALSE,
  boot_iter = 10,
  id = "id"
)
}
\arguments{
\item{data}{A data frame containing the data}

\item{time_start}{start time}

\item{time_end}{end time}

\item{status}{survival status. 1 for death and 0 for censored}

\item{exposure}{Name of the time-varying exposure variable. From the most recent to the furthest in time.}

\item{knots_number}{A vector of prespecified knot number}

\item{latency}{prespecified latency}

\item{adjusted_variable}{A vector of adjusted variables}

\item{adjusted_model}{A vector of adjusted models}

\item{lag}{A value of lag time}

\item{parallel}{A boolean numeric to indicate whether to run the function in parallel}

\item{boot_iter}{Number of bootstrap iterations}

\item{id}{A string of the column name that contains the subject ID}
}
\value{
A data.frame. The first column is the mean of log HR. The second column is the variance of log HR
}
\description{
This function conduct bootstraps for @CoxKnotsearch
}
\examples{
time_start <- "age_start"
time_end <- "age_end"
status <- "failure"
exposure <- paste0("lag", 0:15)
knots_number <- c(3,4,5)
latency <- 16
adjusted_variable <- "L"
adjusted_model <- "L"
lag <- 2
boot_iter <- 10
id <- "id"
parallel <- TRUE
result <- CoxKnotsearch_boot(sim_data, time_start, time_end, status, exposure, knots_number = knots_number, latency,
lag = lag, parallel = parallel, boot_iter = boot_iter, id = id)
}
