% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxTermsearch.R
\name{CoxTermsearch}
\alias{CoxTermsearch}
\title{This function implments Stepwise model selection by terms of natural cubic splines in Cox regression for latency analysis
This is GreadSearching by knots}
\usage{
CoxTermsearch(
  data,
  time_start,
  time_end,
  status,
  exposure,
  knots,
  latency,
  adjusted_variable = NULL,
  adjusted_model = NULL,
  criteria = "AIC"
)
}
\arguments{
\item{data}{A data frame containing the data}

\item{time_start}{start time}

\item{time_end}{end time}

\item{status}{survival status. 1 for death and 0 for censored}

\item{exposure}{Name of the time-varying exposure variable. From the most recent to the furthest in time.}

\item{knots}{A vector of prespecified knots}

\item{latency}{prespecified latency}

\item{adjusted_variable}{A vector of adjusted variables}

\item{adjusted_model}{A vector of adjusted models}

\item{criteria}{A string to specify the criteria for model selection. It can be either "AIC" or "BIC". The default is "AIC"}
}
\value{
A list. The first element is the best knots with the lowest AIC. The second element is a Cox model
}
\description{
This function implments Stepwise model selection by terms of natural cubic splines in Cox regression for latency analysis
This is GreadSearching by knots
}
\examples{
time_start <- "age_start"
time_end <- "age_end"
status <- "failure"
exposure <- paste0("lag", 0:15)
knots <- seq(0, 15, by = 3)
latency <- 16
adjusted_variable <- "L"
adjusted_model <- "L"
fit <- CoxTermsearch(sim_data, time_start, time_end, status, exposure, knots = knots, latency)
#In this example, we will generate a random covariate L and adjust for it in our model
sim_data$L <- rbinom(nrow(sim_data), 1, 0.5)
adjusted_variable <- "L"
adjusted_model <- "L+I(L^2)"
fit <- CoxTermsearch(sim_data, time_start, time_end, status, exposure, knots = knots, latency,
adjusted_variable = adjusted_variable, adjusted_model = adjusted_model)
#L is selected into the final model but L^2 was not
}
