% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxTermsearch.R
\name{CoxTermsearch_boot}
\alias{CoxTermsearch_boot}
\title{This function conduct bootstraps for @CoxTermsearch}
\usage{
CoxTermsearch_boot(
  data,
  time_start,
  time_end,
  status,
  exposure,
  knots,
  latency,
  adjusted_variable = NULL,
  adjusted_model = NULL,
  lag,
  parallel = FALSE,
  boot_iter = 10,
  id = "id",
  criteria = "AIC"
)
}
\arguments{
\item{data}{A data frame containing the data}

\item{time_start}{start time}

\item{time_end}{end time}

\item{status}{survival status. 1 for death and 0 for censored}

\item{exposure}{Name of the time-varying exposure variable. From the most recent to the furthest in time.}

\item{knots}{A vector of prespecified knots}

\item{latency}{prespecified latency}

\item{adjusted_variable}{A vector of adjusted variables}

\item{adjusted_model}{A vector of adjusted models}

\item{lag}{A value of lag time}

\item{parallel}{A boolean numeric to indicate whether to run the function in parallel}

\item{boot_iter}{Number of bootstrap iterations}

\item{id}{A string of the column name that contains the subject ID}

\item{criteria}{A string to specify the criteria for model selection. It can be either "AIC" or "BIC". The default is "AIC"}
}
\value{
A data.frame. The first column is the mean of log HR. The second column is the variance of log HR, the third and
fourth columns are the 2.5 and 97.5 percentiles of log HR if parallel = TRUE
}
\description{
This function conduct bootstraps for @CoxTermsearch
}
\examples{
time_start <- "age_start"
time_end <- "age_end"
status <- "failure"
exposure <- paste0("lag", 0:15)
knots <- seq(0, 15, by = 3)
latency <- 16
adjusted_variable <- "L"
adjusted_model <- "L"
lag <- 2
boot_iter <- 1
id <- "id"
parallel <- TRUE
result <- CoxTermsearch_boot(sim_data, time_start, time_end, 
status, exposure, knots = knots, latency,
lag = lag, parallel = parallel, boot_iter = boot_iter, id = id)
}
