% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxNCSpline.R
\name{extract_CoxNCSpline}
\alias{extract_CoxNCSpline}
\title{This is a function to extract log HR based on the resuls from CoxNCSpline}
\usage{
extract_CoxNCSpline(fit, lag, latency)
}
\arguments{
\item{fit}{A model from CoxNCSpline}

\item{lag}{A value of lag time}

\item{latency}{prespecified latency}
}
\value{
The log HR estimated at that lag time
}
\description{
This is a function to extract log HR based on the resuls from CoxNCSpline
}
\examples{
time_start <- "age_start"
time_end <- "age_end"
status <- "failure"
exposure <- paste0("lag", 0:19)
knots_number <- 3
latency <- 20
fit <- CoxNCSpline(sim_data, time_start, time_end, status, exposure, knots_number, latency)
extract_CoxNCSpline(fit, 10, latency = latency)
}
