% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxPoly.R
\name{extract_CoxPoly}
\alias{extract_CoxPoly}
\title{This is a function to extract log HR based on the resuls from CoxPoly}
\usage{
extract_CoxPoly(fit, lag)
}
\arguments{
\item{fit}{A model from CoxPoly}

\item{lag}{A value of lag time}
}
\value{
The log HR estimated at that lag time
}
\description{
This is a function to extract log HR based on the resuls from CoxPoly
}
\examples{
time_start <- "age_start"
time_end <- "age_end"
status <- "failure"
exposure <- paste0("lag", 0:19)
degree <- 3
latency <- 20
fit <- CoxPoly(sim_data, time_start, time_end, status, exposure, degree, latency)
extract_CoxPoly(fit, 10)
}
